+++ src/LogEntry/LogEntry.php
@@ -14,6 +14,11 @@ class LogEntry
     protected $type;
 
     /**
+     * @var string
+     */
+    protected $scope;
+
+    /**
      * The log entry audience, e.g.: *|dev|user
      * @var $audience
      */
@@ -41,15 +46,19 @@ class LogEntry
 
     /**
      * LogEntry constructor.
-     * @param string $type
+     * @param string $ccMessage
      * @param string $audience
      * @param \DateTimeImmutable|null $date
-     * @param null $desc
      * @throws LogEntryException
      */
-    public function __construct($type = 'feat', $audience = '*', \DateTimeImmutable $date = null, $desc = null)
+    public function __construct($ccMessage = 'misc: progress', $audience = '*', \DateTimeImmutable $date = null)
     {
-        $this->type = LogEntryType::getCanonicalIdentifier($type);
+        $ccMessageElements = static::parseElementsFromCcMessage($ccMessage);
+
+        $this->type = $ccMessageElements['type'];
+        $this->scope = $ccMessageElements['scope'];
+        $this->desc = $ccMessageElements['desc'];
+
         $this->audience = $audience;
 
         if (!$date) {
@@ -57,14 +66,15 @@ class LogEntry
         }
 
         $this->date = $date;
-        $this->desc = $desc;
     }
 
     /**
      * @param array $entryArr
      * @throws LogEntryException
+     *
+     * TODO: Extract as factory method
      */
-    public function parseFromArray(array $entryArr)
+    public function createFromArray(array $entryArr)
     {
         $mandatoryKeys = ['date', 'type', 'desc'];
 
@@ -78,7 +88,12 @@ class LogEntry
 
             switch ($key) {
                 case 'type':
-                    $this->setType(LogEntryType::getCanonicalIdentifier(strtolower($value)));
+                    $this->setTypeByCcType($value);
+                    break;
+
+                case 'scope':
+                    // Explicitly passed scopes override scopes parsed from type
+                    $this->setScope($value);
                     break;
 
                 case 'date':
@@ -86,7 +101,6 @@ class LogEntry
                     break;
 
                 case 'desc':
-                case 'message':
                     $this->setDesc($value);
                     break;
 
@@ -107,6 +121,82 @@ class LogEntry
     }
 
     /**
+     * Parse elements (type, scope, description) from "conventional commit" message.
+     * Message string missing type and optional scope are handled gracefully so that
+     * an array with whole message string as value for key "desc" is returned.
+     * Messages of the following format will be matched:
+     *
+     * - type: description
+     * - [type]: description
+     * - type(scope): description
+     * - type (scope): description
+     * - [type(scope)]: description
+     * - [type (scope)]: description
+     *
+     * @param $message
+     * @param bool $canonical
+     * @return array
+     * @throws LogEntryException
+     */
+    public static function parseElementsFromCcMessage($message, $canonical = true)
+    {
+        $res = [
+            'type' => 'misc',
+            'scope' => null,
+            'desc' => $message
+        ];
+
+        $ccTypePatterns = implode('|', LogEntryType::getUniqueTypeAliases());
+
+        $regex = '/^(\[?\b(' . $ccTypePatterns . ')\b[\s]?(\((\b.*\b)\))?\]?\:[\s]?)(.*)$/si';
+
+        if (preg_match($regex, $message, $matches)) {
+            $res = [
+                'type' => $canonical ? LogEntryType::getCanonicalIdentifier($matches[2]) : $matches[2],
+                'scope' => $matches[4],
+                'desc' => $matches[5]
+            ];
+        }
+
+        return $res;
+    }
+
+    /**
+     * Parse elements (type, scope) from "conventional commit" type.
+     * Type string missing scope are handled gracefully so that
+     * an array with whole type string as value for key "type" is returned.
+     * Types of the following format will be matched:
+     *
+     * - type
+     * - type(scope)
+     *
+     * @param $type
+     * @param bool $canonical
+     * @return array
+     * @throws LogEntryException
+     */
+    public static function parseElementsFromCcType($type, $canonical = true)
+    {
+        $ccTypePatterns = implode('|', LogEntryType::getUniqueTypeAliases());
+
+        $regex = '/^(\b' . $ccTypePatterns . '\b)[\s]?\((\b.*\b)\)$/si';
+
+        if (preg_match($regex, strtolower($type), $matches)) {
+            $res = [
+                'type' => $canonical ? LogEntryType::getCanonicalIdentifier($matches[1]) : $matches[1],
+                'scope' => $matches[2]
+            ];
+        } else {
+            $res = [
+                'type' => LogEntryType::getCanonicalIdentifier($type),
+                'scope' => null
+            ];
+        }
+
+        return $res;
+    }
+
+    /**
      * @return mixed
      */
     public function getType()
@@ -115,11 +205,39 @@ class LogEntry
     }
 
     /**
-     * @param mixed $type
+     * Set type and scope by parsing "conventional commit" style type
+     * @param $ccType
+     * @throws LogEntryException
+     */
+    public function setTypeByCcType($ccType)
+    {
+        $ccTypeElements = static::parseElementsFromCcType($ccType);
+        $this->setType($ccTypeElements['type']);
+        $this->setScope($ccTypeElements['scope']);
+    }
+
+    /**
+     * @param $type
      */
     public function setType($type)
     {
         $this->type = $type;
+    }
+
+    /**
+     * @return string
+     */
+    public function getScope()
+    {
+        return $this->scope;
+    }
+
+    /**
+     * @param string $scope
+     */
+    public function setScope($scope)
+    {
+        $this->scope = $scope;
     }
 
     /**